squash! csky: Emulate one-byte and two-byte cmpxchg
[ paulmck: Apply kernel test robot feedback. ]

+++ arch/csky/include/asm/cmpxchg.h
@@ -62,10 +62,10 @@
 	__typeof__(*(ptr)) __ret;				\
 	switch (size) {						\
 	case 1:							\
-		__ret = cmpxchg_emu_u8((volatile u8 *)__ptr, __old, __new); \
+		__ret = cmpxchg_emu_u8((volatile u8 *)__ptr, (uintptr_t)__old, (uintptr_t)__new); \
 		break;						\
 	case 2:							\
-		__ret = cmpxchg_emu_u16((volatile u16 *)__ptr, __old, __new); \
+		__ret = cmpxchg_emu_u16((volatile u16 *)__ptr, (uintptr_t)__old, (uintptr_t)__new); \
 		break;						\
 	case 4:							\
 		asm volatile (					\
@@ -98,10 +98,10 @@
 	__typeof__(*(ptr)) __ret;				\
 	switch (size) {						\
 	case 1:							\
-		__ret = cmpxchg_emu_u8((volatile u8 *)__ptr, __old, __new); \
+		__ret = cmpxchg_emu_u8((volatile u8 *)__ptr, (uintptr_t)__old, (uintptr_t)__new); \
 		break;						\
 	case 2:							\
-		__ret = cmpxchg_emu_u16((volatile u16 *)__ptr, __old, __new); \
+		__ret = cmpxchg_emu_u16((volatile u16 *)__ptr, (uintptr_t)__old, (uintptr_t)__new); \
 		break;						\
 	case 4:							\
 		asm volatile (					\
@@ -135,10 +135,10 @@
 	__typeof__(*(ptr)) __ret;				\
 	switch (size) {						\
 	case 1:							\
-		__ret = cmpxchg_emu_u8((volatile u8 *)__ptr, __old, __new); \
+		__ret = cmpxchg_emu_u8((volatile u8 *)__ptr, (uintptr_t)__old, (uintptr_t)__new); \
 		break;						\
 	case 2:							\
-		__ret = cmpxchg_emu_u16((volatile u16 *)__ptr, __old, __new); \
+		__ret = cmpxchg_emu_u16((volatile u16 *)__ptr, (uintptr_t)__old, (uintptr_t)__new); \
 		break;						\
 	case 4:							\
 		asm volatile (					\
